# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WFoods ComandaFácil API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (super-admin)
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant/restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant by ID
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # Categories
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for current tenant
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: available
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Public menu (no auth)
  /menu/{tenantSlug}:
    get:
      operationId: getPublicMenu
      tags: [products]
      summary: Get public menu for a restaurant by slug
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicMenu"
        "404":
          description: Restaurant not found

  # Tables
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List tables for current tenant
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "201":
          description: Table created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{tableId}:
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update table
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableBody"
      responses:
        "200":
          description: Table updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete table
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
...
  /waiters:
    get:
      operationId: listWaiters
      tags: [auth]
      summary: List waiters for current tenant
      responses:
        "200":
          description: List of waiters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Waiter"
    post:
      operationId: createWaiter
      tags: [auth]
      summary: Create a waiter for current tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWaiterBody"
      responses:
        "201":
          description: Waiter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Waiter"
  /waiters/{waiterId}:
    delete:
      operationId: deleteWaiter
      tags: [auth]
      summary: Delete waiter
      parameters:
        - name: waiterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - active
        - createdAt
    CreateTenantBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        adminEmail:
          type: string
        adminPassword:
          type: string
      required:
        - name
        - slug
        - adminEmail
        - adminPassword
    UpdateTenantBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
    Category:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - name
        - sortOrder
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        available:
          type: boolean
        printSector:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - name
        - price
        - available
        - createdAt
    CreateProductBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        available:
          type: boolean
        printSector:
          type: ["string", "null"]
      required:
        - name
        - price
    UpdateProductBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        available:
          type: ["boolean", "null"]
        printSector:
          type: ["string", "null"]
    PublicMenu:
      type: object
      properties:
        tenant:
          $ref: "#/components/schemas/Tenant"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required:
        - tenant
        - categories
        - products
    Table:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        number:
          type: integer
        label:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - number
        - active
        - createdAt
    CreateTableBody:
      type: object
      properties:
        number:
          type: integer
        label:
          type: ["string", "null"]
      required:
        - number
    UpdateTableBody:
      type: object
      properties:
        number:
          type: ["integer", "null"]
        label:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]
...
    Waiter:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - name
        - createdAt
    CreateWaiterBody:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
      required:
        - name
        - password