# lib/api-zod/src/generated/types/updateTableBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WFoods ComandaFácil API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTableBody {
  /** @nullable */
  number?: number | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  active?: boolean | null;
}